+++ tidy3d/__init__.py
@@ -47,7 +47,7 @@ from .components import ModeData, ModeAmpsData, ModeIndexData
 from .constants import C_0, ETA_0, HBAR, EPSILON_0, MU_0, Q_e, inf
 
 # plugins typically imported as `from tidy3d.plugins import DispersionFitter`
-from . import plugins
+# from . import plugins
 
 # material library dict imported as `from tidy3d import material_library`
 # get material `mat` and variant `var` as `material_library[mat][var]`